+++ internal/constant/jwt.const.go
@@ -1,8 +1,0 @@
-package constant
-
-import "time"
-
-const (
-	JwtAccessTokenExpirationTime  = 10 * time.Minute
-	JwtRefreshTokenExpirationTime = 6 * time.Hour
-)