+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-# Boilerplate Service
+# Authorization Service
+
+
+<a name="v1.1.0"></a>
+## [v1.1.0] - 2023-04-23
+### New Features
+- cache token when register new user customer
+- impement assign customer role when register
 
 
 <a name="v1.0.0"></a>
@@ -7,4 +14,5 @@
 - init usecase, repository, and grpc transport
 
 
-[Unreleased]: https://github.com/bagastri07/boilerplate-service/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/bagastri07/boilerplate-service/compare/v1.1.0...HEAD
+[v1.1.0]: https://github.com/bagastri07/boilerplate-service/compare/v1.0.0...v1.1.0